+++ src/CreditCardStrategy.java
@@ -8,7 +8,7 @@ public class CreditCardStrategy implements PaymentStrategy {
 	}
 	
 	@Override
-	public void pay(int amount) throws InterruptedException {
+	public void pay(float amount) throws InterruptedException {
 		System.out.println("Inserire cvv per la verifica");
 		@SuppressWarnings("resource")
 		Scanner scanner = new Scanner(System.in);
@@ -16,12 +16,12 @@ public class CreditCardStrategy implements PaymentStrategy {
 		if(acq.getCvv() == code) {
 			System.out.println("Carta "+acq.getCardNumber()+" accettata, tentativo di pagamento in corso...");
 			Thread.sleep(3000);
-			int budget = acq.getBudget();
+			float budget = acq.getBudget();
 			budget -= amount;
 			System.out.println("Pagamento di "+amount+" €, effettuato dal Sig. "+acq.getNome()+" riuscito, il budget rimanente è di "+budget+" €");
 			acq.setBudget(budget);
 		}
 		else
-			System.out.println("Il codice cvv non è corretto");
+			System.out.println("Il codice cvv non è corretto"); // TODO migliora controllo cvv
 	}
 }
