+++ src/SF90StradaleDeluxe.java
@@ -2,12 +2,12 @@ public class SF90StradaleDeluxe implements SF90Stradale{
 	private int hp;
 	private boolean satNav;
 	private boolean leatherWheel;
-	private int price;
+	private float price;
 	private Acquirente acq;
 	private PaymentStrategy method;
 	private Catalogo catalogo;
 	
-	public SF90StradaleDeluxe(int hp, boolean satNav, boolean leatherWheel, int price, Acquirente acq, PaymentStrategy method, Catalogo catalogo) {
+	public SF90StradaleDeluxe(int hp, boolean satNav, boolean leatherWheel, float price, Acquirente acq, PaymentStrategy method, Catalogo catalogo) {
 		this.hp = hp;
 		this.satNav = satNav;
 		this.leatherWheel = leatherWheel;
@@ -19,7 +19,7 @@ public class SF90StradaleDeluxe implements SF90Stradale{
 
 	
 	private void paga(PaymentStrategy paymentMethod) throws InterruptedException {
-		int amount = this.getPrice();
+		float amount = this.getPrice();
 		paymentMethod.pay(amount);
 	}
 	
@@ -66,7 +66,7 @@ public class SF90StradaleDeluxe implements SF90Stradale{
 	public void setLeatherWheel(boolean leatherWheel) {
 		this.leatherWheel = leatherWheel;
 	}
-	public int getPrice() {
+	public float getPrice() {
 		return price;
 	}
 	public Acquirente getAcquirente() {
