+++ src/TransferStrategy.java
@@ -6,9 +6,9 @@ public class TransferStrategy implements PaymentStrategy {
 	}
 
 	@Override
-	public void pay(int amount) throws InterruptedException {
+	public void pay(float amount) throws InterruptedException {
 		System.out.println("Tentativo di pagamento tramite bonifico in corso...");
-		int budget = acq.getBudget();
+		float budget = acq.getBudget();
 		budget -= amount;
 		Thread.sleep(3000);
 		System.out.println(amount+" pagato con bonifico effettuato da "+acq.getNome()+" all'IBAN "+acq.getIBAN()+", rimangono "+budget+" €");
